mars_x/game/game_world.py
import sdl2
from mars_x.cython_modules import physics

class GameWorld:
    def __init__(self, renderer):
        self.renderer = renderer
        self.entities = []
        self.player = None
        self.init_world()
    
    def init_world(self):
        # Initialize game world
        # In a real implementation, this would:
        # - Load level data
        # - Create entities 
        # - Set up the player
        pass
    
    def process_input(self, event):
        # Process input events for the game
        if event.type == sdl2.SDL_KEYDOWN:
            if event.key.keysym.sym == sdl2.SDLK_ESCAPE:
                # Handle escape key
                pass
            elif event.key.keysym.sym == sdl2.SDLK_w:
                # Move forward
                pass
            # Handle other keys...
    
    def update(self):
        # Update game state
        # - Update entity positions using Cython physics
        # - Handle collisions
        # - Update game mechanics
        physics.update_positions(self.entities)
        
        # Process game logic
        for entity in self.entities:
            entity.update()
    
    def render(self):
        # Render the game world
        # In a full implementation, this would use the Vulkan renderer
        # to draw all entities, backgrounds, effects, etc.
        pass
