pyproject.toml
[build-system]
requires = ["setuptools>=64.0", "wheel", "cython>=3.0.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mars-x"
version = "0.1.0"
description = "Mars-X: A 2D top-down space flight game"
authors = [
    {name = "3nigma"}
]
requires-python = ">=3.12"
dependencies = [
    "pysdl2>=0.9.14",
    "pysdl2-dll>=2.26.0",
    "vulkan>=1.3.236",
    "appdirs>=1.4.4"
]

[project.optional-dependencies]
dev = [
    "cython>=3.0.0",
    "setuptools>=68.0.0",
    "wheel>=0.41.0",
    "ninja>=1.11.1"
]

# GUI dependencies - uncomment when needed
# gui = [
#    "glfw>=2.6.2"
# ]

[tool.setuptools]
packages = ["mars_x"]

[project.scripts]
mars-x = "mars_x.main:main"
